+++ decidim-meetings/app/forms/decidim/meetings/admin/meeting_form.rb
@@ -30,7 +30,9 @@ module Decidim
         validates :current_feature, presence: true
         validates :category, presence: true, if: ->(form) { form.decidim_category_id.present? }
         validates :scope, presence: true, if: ->(form) { form.decidim_scope_id.present? }
-        validate { errors.add(:decidim_scope_id, :invalid) if current_participatory_space&.scope && !current_participatory_space&.scope&.ancestor_of?(scope) }
+        validate do
+          errors.add(:decidim_scope_id, :invalid) unless valid_scope?
+        end
 
         delegate :categories, to: :current_feature
 
@@ -47,19 +49,20 @@ module Decidim
         # Returns a Decidim::Scope
         def scope
           return unless current_feature
-          @scope ||= @decidim_scope_id ? current_feature.scopes.find_by(id: @decidim_scope_id) : current_participatory_space&.scope
-        end
-
-        # Scope identifier
-        #
-        # Returns the scope identifier related to the meeting
-        def decidim_scope_id
-          @decidim_scope_id || scope&.id
+          @scope ||= decidim_scope_id ? current_feature.scopes.find_by(id: decidim_scope_id) : current_participatory_space.scope
         end
 
         def category
           return unless current_feature
           @category ||= categories.find_by(id: decidim_category_id)
+        end
+
+        private
+
+        def valid_scope?
+          return true if current_participatory_space.scope.nil?
+
+          current_participatory_space&.scope == scope || current_participatory_space.scope.ancestor_of?(scope)
         end
       end
     end